+++ website/src/components/index.js
@@ -1,9 +1,7 @@
 export { Button } from './Button';
 export { Container } from './Container';
 export { Header } from './Header';
-export { Footer } from './Footer';
 export { SocialIconsNav } from './SocialIconsNav';
 export { Sidebar } from './Sidebar';
-export { SidebarNav } from './SidebarNav';
 export { Search } from './Search';
 export { SiteMeta } from './SiteMeta';
