+++ website/src/components/markdown/Heading.js
@@ -1,6 +1,5 @@
 /** @jsx jsx */
 
-import React from 'react'; // eslint-disable-line no-unused-vars
 import { jsx } from '@emotion/core';
 import { colors, gridSize } from '@arch-ui/theme';
 import { LinkIcon } from '@arch-ui/icons';
@@ -43,9 +42,11 @@ const Heading = ({ as: Tag, children, id, ...props }) => {
     <Tag
       css={{
         color: colors.N100,
+        fontWeight: 700,
+        letterSpacing: '-0.025em',
         lineHeight: 1,
         marginBottom: '0.66em',
-        marginTop: 0,
+        marginTop: '1.66em',
       }}
       id={id}
       {...props}
@@ -71,7 +72,7 @@ export const H1 = props => (
   <Heading
     css={mq({
       fontSize: '2.4rem',
-      fontWeight: 300,
+      fontWeight: 500,
       marginTop: 0,
     })}
     {...props}
@@ -83,6 +84,7 @@ export const H2 = props => (
     {...props}
     css={mq({
       fontSize: '1.8rem',
+      fontWeight: 500,
 
       // offset header height for hash links
       '::before': {
@@ -100,14 +102,13 @@ export const H3 = props => (
     css={mq({
       fontSize: '1.4rem',
       fontWeight: 500,
-      paddingTop: '1.5em',
     })}
     {...props}
     as="h3"
   />
 );
-export const H4 = props => <Heading css={{ fontSize: '1.2rem' }} {...props} as="h4" />;
-export const H5 = props => <Heading {...props} css={{ fontSize: '1rem' }} as="h5" />;
-export const H6 = props => <Heading {...props} css={{ fontSize: '0.9rem' }} as="h6" />;
+export const H4 = props => <Heading css={{ fontSize: '1.2rem' }} as="h4" {...props} />;
+export const H5 = props => <Heading css={{ fontSize: '1rem' }} as="h5" {...props} />;
+export const H6 = props => <Heading css={{ fontSize: '0.9rem' }} as="h6" {...props} />;
 
 export default Heading;
