+++ website/src/components/markdown/mdx-renderer.js
@@ -1,0 +1,3 @@
+export default function PostPageTemplate(props) {
+  return props.children;
+}
