--- website/src/components/mdx-renderer.js
@@ -1,5 +1,0 @@
-function PostPageTemplate(props) {
-  return props.children;
-}
-
-export default PostPageTemplate;
