+++ website/src/prism-themes/custom.js
@@ -3,8 +3,10 @@
 const colors = {
   blue: '#0065FF',
   red: '#DE350B',
+  redLight: '#FFEEF0',
   yellow: '#FF991F',
   green: '#36B37E',
+  greenLight: '#E6FFED',
   teal: '#00B8D9',
   purple: '#6554C0',
   gray: '#5E6C84',
@@ -87,13 +89,13 @@ export default {
     {
       types: ['deleted'],
       style: {
-        textDecorationLine: 'line-through',
+        backgroundColor: colors.redLight,
       },
     },
     {
       types: ['inserted'],
       style: {
-        textDecorationLine: 'underline',
+        backgroundColor: colors.greenLight,
       },
     },
     {
