+++ website/src/templates/docs.js
@@ -1,10 +1,9 @@
 /** @jsx jsx */
 
-import React from 'react'; // eslint-disable-line no-unused-vars
+import { Fragment } from 'react';
 import { Helmet } from 'react-helmet';
 import { Link, graphql } from 'gatsby';
 import { MDXRenderer } from 'gatsby-plugin-mdx';
-
 import { MDXProvider } from '@mdx-js/react';
 import { jsx } from '@emotion/core';
 import { SkipNavContent } from '@reach/skip-nav';
@@ -14,11 +13,11 @@ import slugify from '@sindresorhus/slugify';
 import { Layout, Content } from '../templates/layout';
 import mdComponents from '../components/markdown';
 import { SiteMeta } from '../components/SiteMeta';
-import { media, mediaMax } from '../utils/media';
-import { useNavData } from '../utils/hooks';
-import { titleCase } from '../utils/case';
 import { Container } from '../components';
 import { Sidebar } from '../components/Sidebar';
+import { media } from '../utils/media';
+import { useNavData } from '../utils/hooks';
+import { titleCase } from '../utils/case';
 
 export default function Template({
   data: { mdx, site }, // this prop will be injected by the GraphQL query below.
@@ -47,7 +46,7 @@ export default function Template({
   }${suffix}`;
 
   return (
-    <>
+    <Fragment>
       <SiteMeta pathname={fields.slug} />
       <Helmet>
         <title>{title}</title>
@@ -59,42 +58,18 @@ export default function Template({
         <meta name="twitter:description" content={fields.description} />
       </Helmet>
       <Layout>
-        {({ sidebarOffset, sidebarIsVisible, toggleSidebar }) => (
+        {({ sidebarIsVisible, toggleSidebar }) => (
           <Container hasGutters={false} css={{ display: 'flex' }}>
-            <Sidebar
-              isVisible={sidebarIsVisible}
-              offsetTop={sidebarOffset}
-              toggleSidebar={toggleSidebar}
-            />
-            <Content className="docSearch-content" css={{ display: 'flex' }}>
+            <Sidebar isVisible={sidebarIsVisible} toggleSidebar={toggleSidebar} />
+            <Content
+              className="docSearch-content"
+              css={{ alignItems: 'flex-start', display: 'flex', flex: 1 }}
+            >
               <div css={{ flex: 1, minWidth: 0 }}>
                 <SkipNavContent />
                 <MDXProvider components={mdComponents}>
                   <MDXRenderer>{body}</MDXRenderer>
                 </MDXProvider>
-                <EditSection>
-                  <p>
-                    Have you found a mistake, something that is missing, or could be improved on
-                    this page? Please edit the Markdown file on GitHub and submit a PR with your
-                    changes.
-                  </p>
-                  <EditButton
-                    href={fields.editUrl}
-                    target="_blank"
-                    title="Edit this page on GitHub"
-                  >
-                    <svg
-                      fill="currentColor"
-                      height="1.25em"
-                      width="1.25em"
-                      viewBox="0 0 40 40"
-                      css={{ marginLeft: -(gridSize / 2), marginRight: '0.5em' }}
-                    >
-                      <path d="m34.5 11.7l-3 3.1-6.3-6.3 3.1-3q0.5-0.5 1.2-0.5t1.1 0.5l3.9 3.9q0.5 0.4 0.5 1.1t-0.5 1.2z m-29.5 17.1l18.4-18.5 6.3 6.3-18.4 18.4h-6.3v-6.2z" />
-                    </svg>
-                    Edit Page
-                  </EditButton>
-                </EditSection>
                 <Pagination aria-label="Pagination">
                   {prev ? (
                     <PaginationButton to={prev.path}>
@@ -114,11 +89,11 @@ export default function Template({
                   )}
                 </Pagination>
               </div>
+              {/* Table of Contents */}
               <div
                 css={{
                   display: 'none',
                   flexShrink: 0,
-                  maxHeight: 'calc(100vh - 124px)',
                   paddingLeft: gridSize * 4,
                   paddingRight: gridSize * 3,
                   position: 'sticky',
@@ -128,7 +103,17 @@ export default function Template({
                   [media.sm]: { display: 'block' },
                 }}
               >
-                <h4 css={{ marginTop: 0 }}>On this page</h4>
+                <h4
+                  css={{
+                    color: colors.N40,
+                    fontSize: '0.8rem',
+                    fontWeight: 700,
+                    marginTop: 0,
+                    textTransform: 'uppercase',
+                  }}
+                >
+                  On this page
+                </h4>
                 <ul css={{ listStyle: 'none', margin: 0, padding: 0 }}>
                   {headings.map(h => (
                     <li
@@ -142,7 +127,12 @@ export default function Template({
                     >
                       <a
                         css={{
-                          color: colors.N60,
+                          color: colors.N80,
+                          fontSize: '0.9rem',
+
+                          ':hover': {
+                            color: colors.N100,
+                          },
                         }}
                         href={`#${slugify(h.value, { decamelize: false })}`}
                       >
@@ -151,12 +141,31 @@ export default function Template({
                     </li>
                   ))}
                 </ul>
+                <EditSection>
+                  {/* <p>
+                    Have you found a mistake, something that is missing, or could be improved on
+                    this page? Please edit the Markdown file on GitHub and submit a PR with your
+                    changes.
+                  </p> */}
+                  <EditButton href={fields.editUrl} target="_blank">
+                    <svg
+                      fill="currentColor"
+                      height="1.25em"
+                      width="1.25em"
+                      viewBox="0 0 40 40"
+                      css={{ marginLeft: -(gridSize / 2), marginRight: '0.5em' }}
+                    >
+                      <path d="m34.5 11.7l-3 3.1-6.3-6.3 3.1-3q0.5-0.5 1.2-0.5t1.1 0.5l3.9 3.9q0.5 0.4 0.5 1.1t-0.5 1.2z m-29.5 17.1l18.4-18.5 6.3 6.3-18.4 18.4h-6.3v-6.2z" />
+                    </svg>
+                    Edit on GitHub
+                  </EditButton>
+                </EditSection>
               </div>
             </Content>
           </Container>
         )}
       </Layout>
-    </>
+    </Fragment>
   );
 }
 
@@ -201,20 +210,9 @@ const Button = ({ as: Tag, ...props }) => (
 const EditSection = props => (
   <section
     css={{
-      borderBottom: `1px solid ${colors.N10}`,
       borderTop: `1px solid ${colors.N10}`,
-      marginBottom: '3rem',
-      marginTop: '3rem',
-      paddingBottom: '3rem',
-      paddingTop: '3rem',
-
-      p: {
-        marginTop: 0,
-      },
-
-      [mediaMax.sm]: {
-        display: 'none',
-      },
+      marginTop: '1.66rem',
+      paddingTop: '1.66rem',
     }}
     {...props}
   />
